security commit

+++ src/main/java/com/spring/SurveyStorage/service/QuestService.java
@@ -6,8 +6,10 @@ import java.util.List;
 
 public interface QuestService {
      List<Question> getAllQuestions();
-     List<Question> getDetQuestion(int survey_id);
-     void saveQuestion(int survey_id, Question question);
+
+    List<Question> getDetQuestion(long survey_id);
+
+    void saveQuestion(int survey_id, Question question);
      Question getQuest(int id);
 
     void deleteQuest(int id);
